added changes from socket io branch that embed user id in token

+++ server/api/submission_handler.py
@@ -8,8 +8,9 @@ from datetime import datetime
 from api.payment_handler import send_transfer, charge_payment
 from sqlalchemy.exc import DataError, IntegrityError
 from botocore.exceptions import BotoCoreError, ClientError
-from utils import handle_database_error, handle_amazon_error, handle_stripe_error
+from utils import handle_database_error, handle_amazon_error, handle_stripe_error, handle_authentication_error, requires_authentication
 import stripe
+import jwt
 
 submission_handler = Blueprint('submission_handler', __name__)
 
@@ -24,8 +25,9 @@ s3 = boto3.client(
 
 @submission_handler.route('/')
 def show_all(contest_id):
-
     submissions = Submission.query.filter_by(contest_id=contest_id).all()
+    winner = Submission.query.filter_by(
+        contest_id=contest_id, winner=True).first()
 
     submissionsURLs = []
     submissionKeys = []
@@ -40,13 +42,16 @@ def show_all(contest_id):
     except BotoCoreError as e:
         return handle_amazon_error(e, "We could not retrieve the current contest's submissions from the S3 bucket.")
 
-    return jsonify({"files": submissionsURLs, "fileKeys": submissionKeys, "submissionIDs": [submission.id for submission in submissions]})
+    return jsonify({"files": submissionsURLs,
+                    "fileKeys": submissionKeys,
+                    "submissionIDs": [submission.id for submission in submissions],
+                    "winnerID": winner.id if winner else None})
 
 
 @submission_handler.route('/upload', methods=['POST'])
-def upload(contest_id):
+@requires_authentication
+def upload(authenticated_user_id, contest_id):
     file = request.files['file']
-    user_id = 1
 
     try:
         s3_resource = boto3.resource('s3')
@@ -55,7 +60,7 @@ def upload(contest_id):
         my_bucket.Object(file_key).put(Body=file)
 
         submission = Submission(
-            user_id=user_id, contest_id=contest_id, image=file_key)
+            user_id=authenticated_user_id, contest_id=contest_id, image=file_key)
 
         db.session.add(submission)
         db.session.commit()
@@ -77,14 +82,20 @@ def show_submission(contest_id, submission_id):
 
 
 @submission_handler.route('/<int:submission_id>/edit')
-def edit(contest_id, submission_id):
+@requires_authentication
+def edit(authenticated_user_id, contest_id, submission_id):
     submission = Submission.query.get_or_404(submission_id)
+    if submission.artist.id != authenticated_user_id:
+        return handle_authentication_error(user_specific=True)
     return jsonify('Show form for editing this submission', submission.to_dict())
 
 
 @submission_handler.route('/<int:submission_id>', methods=['PATCH', 'PUT'])
-def update(contest_id, submission_id):
+@requires_authentication
+def update(authenticated_user_id, contest_id, submission_id):
     submission = Submission.query.get_or_404(submission_id)
+    if submission.artist.id != authenticated_user_id:
+        return handle_authentication_error(user_specific=True)
     try:
         # if there is no data (request.json is empty) should we return an error?
         if request.json:
@@ -99,8 +110,11 @@ def update(contest_id, submission_id):
 
 
 @submission_handler.route('/<int:submission_id>', methods=['DELETE'])
-def delete(contest_id, submission_id):
+@requires_authentication
+def delete(authenticated_user_id, contest_id, submission_id):
     submission = Submission.query.get_or_404(submission_id)
+    if submission.artist.id != authenticated_user_id:
+        return handle_authentication_error(user_specific=True)
 
     try:
         db.session.delete(submission)
@@ -113,9 +127,11 @@ def delete(contest_id, submission_id):
 
 
 @submission_handler.route('/winner', methods=['POST'])
-def declare_winner(contest_id):
+@requires_authentication
+def declare_winner(authenticated_user_id, contest_id):
     submission = Submission.query.get_or_404(request.json['submission_id'])
-
+    if submission.contest.owner.id != authenticated_user_id:
+        return handle_authentication_error(user_specific=True)
     try:
         submission.winner = True
         charge_payment(contest_id)
